Submit Button Logic Added
Added multiple conditions the the submit button to make sure the guess is valid and that it falls in the range of 0-99. It then checks if the users guess is higher, lower or the exact number. It then displays a corresponding message (currently in the console) depending on what conditions are met.